+++ src/util/constants.ts
@@ -40,7 +40,7 @@ export const messages: Record<LLCode, string> = {
     LLColonForMapping: 'If this colon (:) is for a mapping, consider using \\colon.',
     LLCref: 'Consider using \\cref instead of \\ref.',
     LLDoubleQuotes: 'For double quotes, use ``...\'\'.',
-    LLENDash: 'If these are names, consider using en-dash (--)',
+    LLENDash: 'If these are names, consider using en-dash (--). Otherwise, register as an exception or add a space around "-".',
     LLEqnarray: 'Eqnarray is not recommended. Use align or gather instead.',
     LLNonASCII: 'This is a non-ASCII character. Is this intentional?',
     LLLlGg: 'Consider using \\ll and \\gg instead of << and >>.',